howareyoutwitter/requirements.txt
django
psycopg2-binary
requests
djangorestframework
python-twitter
python-dateutil
django-q
